+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-03-19
+## [Unreleased] - 2022-04-16
 
 ### Added
 
 ### Changed
+* Refactored exchangeBits methods and related of the BitVector class.
 * Replaced calls to methods deprecated in jpt-3.2.0.
 
 ### Deprecated